+++ ethal/ethal/doctype/employee_incentive_bulk/employee_incentive_bulk.py
@@ -10,44 +10,44 @@ import datetime
 from dateutil.relativedelta import relativedelta
 
 class EmployeeIncentiveBulk(Document):
-	def on_submit(self):
-		employees = frappe.db.get_all('Employee Incentive Bulk Detail', {'parent': self.name}, ['employee', 'incentive_hours'], as_list = 1)
-		if employees:
-			for employee in employees:
-				date = self.incentive_date
-				last_day = date + relativedelta(day=1, months=+1, days=-1)
-				first_day = date + relativedelta(day=1)
-				company = frappe.db.get_value('Employee', employee[0], 'company')
+	# def on_submit(self):
+	# 	employees = frappe.db.get_all('Employee Incentive Bulk Detail', {'parent': self.name}, ['employee', 'incentive_hours'], as_list = 1)
+	# 	if employees:
+	# 		for employee in employees:
+	# 			date = self.incentive_date
+	# 			last_day = date + relativedelta(day=1, months=+1, days=-1)
+	# 			first_day = date + relativedelta(day=1)
+	# 			company = frappe.db.get_value('Employee', employee[0], 'company')
 		
-				additional_salary = frappe.db.sql("""
-					select name from `tabAdditional Salary` 
-					where employee = '{0}' and salary_component = '{1}'
-					and company = '{2}' and payroll_date between '{3}' and '{4}' and docstatus = 1 
-					group by name desc limit 1
-				""".format(employee[0], self.salary_component, company, first_day, last_day))
+	# 			additional_salary = frappe.db.sql("""
+	# 				select name from `tabAdditional Salary` 
+	# 				where employee = '{0}' and salary_component = '{1}'
+	# 				and company = '{2}' and payroll_date between '{3}' and '{4}' and docstatus = 1 
+	# 				group by name desc limit 1
+	# 			""".format(employee[0], self.salary_component, company, first_day, last_day))
 				
-				base_value = frappe.get_value('Salary Structure Assignment', {'employee': employee[0], 'docstatus': 1}, 'base')
+	# 			base_value = frappe.get_value('Salary Structure Assignment', {'employee': employee[0], 'docstatus': 1}, 'base')
 				
-				incentive_amount = 0
-				if base_value:
-					incentive_amount = (base_value / 26 / 8) * employee[1]
+	# 			incentive_amount = 0
+	# 			if base_value:
+	# 				incentive_amount = (base_value / 26 / 8) * employee[1]
 
-				incentive_added = 0
-				if additional_salary:
-					incentive_added = frappe.db.get_value('Additional Salary', additional_salary[0][0], 'amount')
+	# 			incentive_added = 0
+	# 			if additional_salary:
+	# 				incentive_added = frappe.db.get_value('Additional Salary', additional_salary[0][0], 'amount')
 					
-					cancel_additional_salary = frappe.get_doc('Additional Salary', additional_salary[0][0])
-					cancel_additional_salary.cancel()
+	# 				cancel_additional_salary = frappe.get_doc('Additional Salary', additional_salary[0][0])
+	# 				cancel_additional_salary.cancel()
 
-				additional_salary = frappe.new_doc('Additional Salary')
-				additional_salary.employee = employee[0]
-				additional_salary.salary_component = self.salary_component
-				additional_salary.amount = incentive_amount	+ incentive_added
-				additional_salary.incentive_amount = incentive_amount
-				additional_salary.payroll_date = self.incentive_date
-				additional_salary.company = company
-				additional_salary.overwrite_salary_structure_amount = 0
-				additional_salary.submit()
+	# 			additional_salary = frappe.new_doc('Additional Salary')
+	# 			additional_salary.employee = employee[0]
+	# 			additional_salary.salary_component = self.salary_component
+	# 			additional_salary.amount = incentive_amount	+ incentive_added
+	# 			additional_salary.incentive_amount = incentive_amount
+	# 			additional_salary.payroll_date = self.incentive_date
+	# 			additional_salary.company = company
+	# 			additional_salary.overwrite_salary_structure_amount = 0
+	# 			additional_salary.submit()
 
 	def get_emp_list(self):
 		"""